+++ assets/js/interactive.js
@@ -68,7 +68,7 @@ export class InteractiveCanvas extends BaseScatterTrack {
       yStart: 2.0, // Start value for y axis
       yEnd: -2.0, // End value for y axis
       step: 0.5, // Step value for drawing ticks along y-axis
-      color: "#000000" // Viz color
+      color: "#000000", // Viz color
     };
 
     // Setup draw canvas
@@ -229,41 +229,51 @@ export class InteractiveCanvas extends BaseScatterTrack {
         });
       }
       // reset dragging behaviour
-      this.markingRegion = false
-      this.drag = false
-    })
+      this.markingRegion = false;
+      this.drag = false;
+    });
     // when user asks for Y zoom
-    this.staticCanvas.addEventListener('zoomY', event => {
+    this.staticCanvas.addEventListener("zoomY", (event) => {
       // clear with old coord context
-      this.clearStaticContent()
+      this.clearStaticContent();
       // also needs to clear for track draw
-      if (event.detail.direction === 'out') {
-        this.log2.yStart += 1
-        this.log2.yEnd -= 1
+      if (event.detail.direction === "out") {
+        this.log2.yStart += 1;
+        this.log2.yEnd -= 1;
         if (this.log2.yStart > 2) {
-          this.log2.step = 1
+          this.log2.step = 1;
         }
       }
-      if (event.detail.direction === 'in' && this.log2.yStart > 1) {
-        this.log2.yStart -= 1
-        this.log2.yEnd += 1
+      if (event.detail.direction === "in" && this.log2.yStart > 1) {
+        this.log2.yStart -= 1;
+        this.log2.yEnd += 1;
         if (this.log2.yStart <= 2) {
-          this.log2.step = 0.5
+          this.log2.step = 0.5;
         }
       }
       // then draw new with new coords
-      this.drawStaticContent()
-      drawTrack({ ...readInputField(), force: true, displayLoading: false, drawTitle: false })
-    })
+      this.drawStaticContent();
+      drawTrack({
+        ...readInputField(),
+        force: true,
+        displayLoading: false,
+        drawTitle: false,
+      });
+    });
   }
 
   // clear static content
-  async clearStaticContent () {
-    const linePadding = 2
-    const staticContext = this.staticCanvas.getContext('2d')
+  async clearStaticContent() {
+    const linePadding = 2;
+    const staticContext = this.staticCanvas.getContext("2d");
     // Fill background colour
-    staticContext.fillStyle = '#F7F9F9'
-    staticContext.fillRect(0, 0, this.staticCanvas.width, this.staticCanvas.height)
+    staticContext.fillStyle = "#F7F9F9";
+    staticContext.fillRect(
+      0,
+      0,
+      this.staticCanvas.width,
+      this.staticCanvas.height,
+    );
     // Make content area visible
     // content window
     staticContext.clearRect(
@@ -273,14 +283,19 @@ export class InteractiveCanvas extends BaseScatterTrack {
       this.staticCanvas.height,
     );
     // area for ticks above content area
-    staticContext.clearRect(0, 0, this.staticCanvas.width, this.y + linePadding)
+    staticContext.clearRect(
+      0,
+      0,
+      this.staticCanvas.width,
+      this.y + linePadding,
+    );
     // Transfer image to visible canvas
-    staticContext.drawImage(this.drawCanvas, 0, 0)
+    staticContext.drawImage(this.drawCanvas, 0, 0);
   }
 
   // Draw static content for interactive canvas
-  async drawStaticContent () {
-    const staticContext = this.staticCanvas.getContext('2d')
+  async drawStaticContent() {
+    const staticContext = this.staticCanvas.getContext("2d");
     // Draw rotated y-axis legends
     drawRotatedText(
       staticContext,
@@ -464,33 +479,38 @@ export class InteractiveCanvas extends BaseScatterTrack {
       .catch((error) => {
         this.allowDraw = true;
 
-      // Transfer image to visible canvas
-      this.blitInteractiveCanvas({ start, end })
-      // Draw chromosome title on the content canvas as a blitting
-      // work around
-      this.titleYPos = result.y_pos - this.titleMargin
-      if (drawTitle) {
-        this.titleBbox !== null && this.blitChromName(
-          { textPosition: this.titleBbox, clearOnly: true })
-        this.titleBbox = this.drawTitle(`Chromosome ${result.chrom}`)
-        this.blitChromName({ textPosition: this.titleBbox })
-      }
+        // Transfer image to visible canvas
+        this.blitInteractiveCanvas({ start, end });
+        // Draw chromosome title on the content canvas as a blitting
+        // work around
+        this.titleYPos = result.y_pos - this.titleMargin;
+        if (drawTitle) {
+          this.titleBbox !== null &&
+            this.blitChromName({
+              textPosition: this.titleBbox,
+              clearOnly: true,
+            });
+          this.titleBbox = this.drawTitle(`Chromosome ${result.chrom}`);
+          this.blitChromName({ textPosition: this.titleBbox });
+        }
 
-      return result
-    }).then((result) => {
-      if (displayLoading) {
-        this.loadingDiv.style.display = 'none'
-      } else {
-        document.getElementsByTagName('body')[0].style.cursor = 'auto'
-      }
-      this.allowDraw = true
-    }).catch(error => {
-      this.allowDraw = true
+        return result;
+      })
+      .then((result) => {
+        if (displayLoading) {
+          this.loadingDiv.style.display = "none";
+        } else {
+          document.getElementsByTagName("body")[0].style.cursor = "auto";
+        }
+        this.allowDraw = true;
+      })
+      .catch((error) => {
+        this.allowDraw = true;
 
-      this.inputField.dispatchEvent(
-        new CustomEvent('error', { detail: { error: error } })
-      )
-    })
+        this.inputField.dispatchEvent(
+          new CustomEvent("error", { detail: { error: error } }),
+        );
+      });
   }
 
   drawTitle(title) {